+++ tetris/tetris/TitleScene.h
@@ -1,0 +1,12 @@
+#pragma once
+
+class Scene;
+
+class TitleScene : public Scene {
+
+    TitleScene();
+    ~TitleScene();
+
+    void draw();
+    void update();
+};
